change name of pages and test package
