+++ src/test/java/com/orangehrm/test/TestTimeFunctionAll.java
@@ -1,13 +1,13 @@
-package test;
+package com.orangehrm.test;
 
 import base.CommonAPI;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import pages.LogInPage;
-import pages.TimeFunctionAll;
-import pages.TimeSheets;
+import com.orangehrm.pages.LogInPage;
+import com.orangehrm.pages.TimeFunctionAll;
+import com.orangehrm.pages.TimeSheets;
 import utility.ReadFromExcel;
