+++ src/test/resources/orangehrm-runner.xml
@@ -11,18 +11,18 @@
     <parameter name="url" value="https://opensource-demo.orangehrmlive.com/"></parameter>
     <test name="orangehrm tests">
         <classes>
-            <class name="test.TestLogIn"></class>
-            <class name="test.TestLogOut"></class>
-            <class name="test.TestAddInfoEverything"></class>
-            <class name="test.TestManagerReviews"></class>
-            <class name="test.TestMostLikePost"></class>
-            <class name="test.TestPostComments"></class>
-            <class name="test.TestPIM"></class>
-            <class name="test.TestQualifications"></class>
-            <class name="test.TestReportsOfQAEngineers"></class>
-            <class name="test.TestTimeFunctionAll"></class>
-            <class name="test.TestTimeSheets"></class>
-            <class name="test.AdminPageTest"></class>
+            <class name="com.orangehrm.test.TestLogIn"></class>
+            <class name="com.orangehrm.test.TestLogOut"></class>
+            <class name="com.orangehrm.test.TestAddInfoEverything"></class>
+            <class name="com.orangehrm.test.TestManagerReviews"></class>
+            <class name="com.orangehrm.test.TestMostLikePost"></class>
+            <class name="com.orangehrm.test.TestPostComments"></class>
+            <class name="com.orangehrm.test.TestPIM"></class>
+            <class name="com.orangehrm.test.TestQualifications"></class>
+            <class name="com.orangehrm.test.TestReportsOfQAEngineers"></class>
+            <class name="com.orangehrm.test.TestTimeFunctionAll"></class>
+            <class name="com.orangehrm.test.TestTimeSheets"></class>
+            <class name="com.orangehrm.test.AdminPageTest"></class>
         </classes>
     </test>
 </suite>
