+++ test-output/junitreports/TEST-test.AdminPageTest.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitReportReporter -->
-<testsuite hostname="KABIR" failures="0" tests="3" name="test.AdminPageTest" time="1.922" errors="1" timestamp="2023-03-08T01:52:57 EST" skipped="2">
-  <testcase classname="test.AdminPageTest" name="AddJobTitle" time="0.000">
+<testsuite hostname="KABIR" failures="0" tests="3" name="com.orangehrm.test.AdminPageTest" time="1.922" errors="1" timestamp="2023-03-08T01:52:57 EST" skipped="2">
+  <testcase classname="com.orangehrm.test.AdminPageTest" name="AddJobTitle" time="0.000">
     <skipped/>
   </testcase> <!-- AddJobTitle -->
   <system-out/>
-  <testcase classname="test.AdminPageTest" name="ChangePasswordCheckBox" time="0.000">
+  <testcase classname="com.orangehrm.test.AdminPageTest" name="ChangePasswordCheckBox" time="0.000">
     <skipped/>
   </testcase> <!-- ChangePasswordCheckBox -->
   <system-out/>
-  <testcase classname="test.AdminPageTest" name="setUp" time="1.922">
+  <testcase classname="com.orangehrm.test.AdminPageTest" name="setUp" time="1.922">
     <error message="Could not start a new session. Response code 500. Message: session not created: This version of ChromeDriver only supports Chrome version 111
 Current browser version is 110.0.5481.178 with binary path C:\Program Files\Google\Chrome\Application\chrome.exe 
 Host info: host: &#039;KABIR&#039;, ip: &#039;192.168.1.158&#039;
@@ -73,4 +73,4 @@ at com.intellij.rt.testng.RemoteTestNGStarter.main(RemoteTestNGStarter.java:109)
     </error>
   </testcase> <!-- setUp -->
   <system-out/>
-</testsuite> <!-- test.AdminPageTest -->
+</testsuite> <!-- com.orangehrm.test.AdminPageTest -->
