+++ test-output/old/orangehrm test/classes.html
@@ -4,7 +4,7 @@
 <th>Method name</th>
 <th>Groups</th>
 </tr><tr>
-<td>test.AdminPageTest</td>
+<td>com.orangehrm.test.AdminPageTest</td>
 <td>&nbsp;</td><td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@Test</td>
