+++ test-output/old/orangehrm test/methods-alphabetical.html
@@ -1,24 +1,24 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>orangehrm test</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="d572c1">  <td>23/03/08 01:52:57</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="AdminPageTest.AddJobTitle()[pri:0, instance:test.AdminPageTest@392a04e7]">AddJobTitle</td> 
+<tr bgcolor="d572c1">  <td>23/03/08 01:52:57</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="AdminPageTest.AddJobTitle()[pri:0, instance:com.orangehrm.test.AdminPageTest@392a04e7]">AddJobTitle</td>
   <td>main@2000530216</td>   <td></td> </tr>
-<tr bgcolor="d572c1">  <td>23/03/08 01:52:57</td>   <td>2</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="AdminPageTest.ChangePasswordCheckBox()[pri:0, instance:test.AdminPageTest@392a04e7]">ChangePasswordCheckBox</td> 
+<tr bgcolor="d572c1">  <td>23/03/08 01:52:57</td>   <td>2</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="AdminPageTest.ChangePasswordCheckBox()[pri:0, instance:com.orangehrm.test.AdminPageTest@392a04e7]">ChangePasswordCheckBox</td>
   <td>main@2000530216</td>   <td></td> </tr>
-<tr bgcolor="71beb3">  <td>23/03/08 01:52:57</td>   <td>3</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;CommonAPI.afterEachTestMethod(org.testng.ITestResult)[pri:0, instance:test.AdminPageTest@392a04e7]">&lt;&lt;afterEachTestMethod</td> 
+<tr bgcolor="71beb3">  <td>23/03/08 01:52:57</td>   <td>3</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;CommonAPI.afterEachTestMethod(org.testng.ITestResult)[pri:0, instance:com.orangehrm.test.AdminPageTest@392a04e7]">&lt;&lt;afterEachTestMethod</td>
 <td>&nbsp;</td>  <td></td>   <td></td> </tr>
-<tr bgcolor="71beb3">  <td>23/03/08 01:52:57</td>   <td>3</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;CommonAPI.afterEachTestMethod(org.testng.ITestResult)[pri:0, instance:test.AdminPageTest@392a04e7]">&lt;&lt;afterEachTestMethod</td> 
+<tr bgcolor="71beb3">  <td>23/03/08 01:52:57</td>   <td>3</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;CommonAPI.afterEachTestMethod(org.testng.ITestResult)[pri:0, instance:com.orangehrm.test.AdminPageTest@392a04e7]">&lt;&lt;afterEachTestMethod</td>
 <td>&nbsp;</td>  <td></td>   <td></td> </tr>
-<tr bgcolor="71beb3">  <td>23/03/08 01:52:55</td>   <td>-2019</td> <td title="&gt;&gt;CommonAPI.extentSetup(org.testng.ITestContext)[pri:0, instance:test.AdminPageTest@392a04e7]">&gt;&gt;extentSetup</td> 
+<tr bgcolor="71beb3">  <td>23/03/08 01:52:55</td>   <td>-2019</td> <td title="&gt;&gt;CommonAPI.extentSetup(org.testng.ITestContext)[pri:0, instance:com.orangehrm.test.AdminPageTest@392a04e7]">&gt;&gt;extentSetup</td>
 <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@2000530216</td>   <td></td> </tr>
-<tr bgcolor="71beb3">  <td>23/03/08 01:52:57</td>   <td>28</td> <td title="&lt;&lt;CommonAPI.generateReport()[pri:0, instance:test.AdminPageTest@392a04e7]">&lt;&lt;generateReport</td> 
+<tr bgcolor="71beb3">  <td>23/03/08 01:52:57</td>   <td>28</td> <td title="&lt;&lt;CommonAPI.generateReport()[pri:0, instance:com.orangehrm.test.AdminPageTest@392a04e7]">&lt;&lt;generateReport</td>
 <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@2000530216</td>   <td></td> </tr>
-<tr bgcolor="71beb3">  <td>23/03/08 01:52:57</td>   <td>2</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;CommonAPI.setUp(boolean,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String)[pri:0, instance:test.AdminPageTest@392a04e7]">&gt;&gt;setUp</td> 
+<tr bgcolor="71beb3">  <td>23/03/08 01:52:57</td>   <td>2</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;CommonAPI.setUp(boolean,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String)[pri:0, instance:com.orangehrm.test.AdminPageTest@392a04e7]">&gt;&gt;setUp</td>
 <td>&nbsp;</td>  <td>main@2000530216</td>   <td></td> </tr>
-<tr bgcolor="71beb3">  <td>23/03/08 01:52:57</td>   <td>2</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;CommonAPI.setUp(boolean,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String)[pri:0, instance:test.AdminPageTest@392a04e7]">&gt;&gt;setUp</td> 
+<tr bgcolor="71beb3">  <td>23/03/08 01:52:57</td>   <td>2</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;CommonAPI.setUp(boolean,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String)[pri:0, instance:com.orangehrm.test.AdminPageTest@392a04e7]">&gt;&gt;setUp</td>
 <td>&nbsp;</td>  <td>main@2000530216</td>   <td></td> </tr>
-<tr bgcolor="71beb3">  <td>23/03/08 01:52:57</td>   <td>2</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;CommonAPI.startExtent(java.lang.reflect.Method)[pri:0, instance:test.AdminPageTest@392a04e7]">&gt;&gt;startExtent</td> 
+<tr bgcolor="71beb3">  <td>23/03/08 01:52:57</td>   <td>2</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;CommonAPI.startExtent(java.lang.reflect.Method)[pri:0, instance:com.orangehrm.test.AdminPageTest@392a04e7]">&gt;&gt;startExtent</td>
 <td>&nbsp;</td>  <td></td>   <td></td> </tr>
-<tr bgcolor="71beb3">  <td>23/03/08 01:52:57</td>   <td>2</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;CommonAPI.startExtent(java.lang.reflect.Method)[pri:0, instance:test.AdminPageTest@392a04e7]">&gt;&gt;startExtent</td> 
+<tr bgcolor="71beb3">  <td>23/03/08 01:52:57</td>   <td>2</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;CommonAPI.startExtent(java.lang.reflect.Method)[pri:0, instance:com.orangehrm.test.AdminPageTest@392a04e7]">&gt;&gt;startExtent</td>
 <td>&nbsp;</td>  <td></td>   <td></td> </tr>
 </table>
