+++ test-output/old/orangehrm test/testng.xml.html
@@ -1,1 +1,1 @@
-<html><head><title>testng.xml for orangehrm test</title></head><body><tt>&lt;?xml&nbsp;version="1.0"&nbsp;encoding="UTF-8"?&gt;<br/>&lt;!DOCTYPE&nbsp;suite&nbsp;SYSTEM&nbsp;"https://testng.org/testng-1.0.dtd"&gt;<br/>&lt;suite&nbsp;name="orangehrm&nbsp;test"&nbsp;guice-stage="DEVELOPMENT"&gt;<br/>&nbsp;&nbsp;&lt;parameter&nbsp;name="useCloudEnv"&nbsp;value="false"/&gt;<br/>&nbsp;&nbsp;&lt;parameter&nbsp;name="os"&nbsp;value="windows"/&gt;<br/>&nbsp;&nbsp;&lt;parameter&nbsp;name="osVersion"&nbsp;value="10"/&gt;<br/>&nbsp;&nbsp;&lt;parameter&nbsp;name="envName"&nbsp;value="browserstack"/&gt;<br/>&nbsp;&nbsp;&lt;parameter&nbsp;name="browserVersion"&nbsp;value="100"/&gt;<br/>&nbsp;&nbsp;&lt;parameter&nbsp;name="browserName"&nbsp;value="Chrome"/&gt;<br/>&nbsp;&nbsp;&lt;parameter&nbsp;name="url"&nbsp;value="https://opensource-demo.orangehrmlive.com/"/&gt;<br/>&nbsp;&nbsp;&lt;test&nbsp;thread-count="5"&nbsp;name="orangehrm&nbsp;tests"&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;classes&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&lt;class&nbsp;name="test.AdminPageTest"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;/classes&gt;<br/>&nbsp;&nbsp;&lt;/test&gt;&nbsp;&lt;!--&nbsp;orangehrm&nbsp;tests&nbsp;--&gt;<br/>&lt;/suite&gt;&nbsp;&lt;!--&nbsp;orangehrm&nbsp;test&nbsp;--&gt;<br/></tt></body></html>
+<html><head><title>testng.xml for orangehrm test</title></head><body><tt>&lt;?xml&nbsp;version="1.0"&nbsp;encoding="UTF-8"?&gt;<br/>&lt;!DOCTYPE&nbsp;suite&nbsp;SYSTEM&nbsp;"https://testng.org/testng-1.0.dtd"&gt;<br/>&lt;suite&nbsp;name="orangehrm&nbsp;test"&nbsp;guice-stage="DEVELOPMENT"&gt;<br/>&nbsp;&nbsp;&lt;parameter&nbsp;name="useCloudEnv"&nbsp;value="false"/&gt;<br/>&nbsp;&nbsp;&lt;parameter&nbsp;name="os"&nbsp;value="windows"/&gt;<br/>&nbsp;&nbsp;&lt;parameter&nbsp;name="osVersion"&nbsp;value="10"/&gt;<br/>&nbsp;&nbsp;&lt;parameter&nbsp;name="envName"&nbsp;value="browserstack"/&gt;<br/>&nbsp;&nbsp;&lt;parameter&nbsp;name="browserVersion"&nbsp;value="100"/&gt;<br/>&nbsp;&nbsp;&lt;parameter&nbsp;name="browserName"&nbsp;value="Chrome"/&gt;<br/>&nbsp;&nbsp;&lt;parameter&nbsp;name="url"&nbsp;value="https://opensource-demo.orangehrmlive.com/"/&gt;<br/>&nbsp;&nbsp;&lt;test&nbsp;thread-count="5"&nbsp;name="orangehrm&nbsp;tests"&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;classes&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&lt;class&nbsp;name="com.orangehrm.test.AdminPageTest"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;/classes&gt;<br/>&nbsp;&nbsp;&lt;/test&gt;&nbsp;&lt;!--&nbsp;orangehrm&nbsp;tests&nbsp;--&gt;<br/>&lt;/suite&gt;&nbsp;&lt;!--&nbsp;orangehrm&nbsp;test&nbsp;--&gt;<br/></tt></body></html>
