+++ test-output/orangehrm test/orangehrm tests.html
@@ -76,7 +76,7 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='test.AdminPageTest.setUp()'><b>setUp</b><br>Test class: test.AdminPageTest<br>Parameters: false, browserstack, windows, 10, Chrome, 100, https://opensource-demo.orangehrmlive.com/</td>
+<td title='com.orangehrm.test.AdminPageTest.setUp()'><b>setUp</b><br>Test class: com.orangehrm.test.AdminPageTest<br>Parameters: false, browserstack, windows, 10, Chrome, 100, https://opensource-demo.orangehrmlive.com/</td>
 <td><div><pre>org.openqa.selenium.SessionNotCreatedException: Could not start a new session. Response code 500. Message: session not created: This version of ChromeDriver only supports Chrome version 111
 Current browser version is 110.0.5481.178 with binary path C:\Program Files\Google\Chrome\Application\chrome.exe 
 Host info: host: &apos;KABIR&apos;, ip: &apos;192.168.1.158&apos;
@@ -156,7 +156,7 @@ Command: [null, newSession {capabilities=[Capabilities {browserName: chrome, goo
 	at com.intellij.rt.testng.RemoteTestNGStarter.main(RemoteTestNGStarter.java:109)
 </pre></div></td>
 <td>1</td>
-<td>test.AdminPageTest@392a04e7</td></tr>
+<td>com.orangehrm.test.AdminPageTest@392a04e7</td></tr>
 </table><p>
 <table width='100%' border='1' class='invocation-skipped'>
 <tr><td colspan='4' align='center'><b>SKIPPED CONFIGURATIONS</b></td></tr>
@@ -167,30 +167,30 @@ Command: [null, newSession {capabilities=[Capabilities {browserName: chrome, goo
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='test.AdminPageTest.setUp()'><b>setUp</b><br>Test class: test.AdminPageTest</td>
+<td title='com.orangehrm.test.AdminPageTest.setUp()'><b>setUp</b><br>Test class: com.orangehrm.test.AdminPageTest</td>
 <td></td>
 <td>0</td>
-<td>test.AdminPageTest@392a04e7</td></tr>
+<td>com.orangehrm.test.AdminPageTest@392a04e7</td></tr>
 <tr>
-<td title='test.AdminPageTest.startExtent()'><b>startExtent</b><br>Test class: test.AdminPageTest</td>
+<td title='com.orangehrm.test.AdminPageTest.startExtent()'><b>startExtent</b><br>Test class: com.orangehrm.test.AdminPageTest</td>
 <td></td>
 <td>0</td>
-<td>test.AdminPageTest@392a04e7</td></tr>
+<td>com.orangehrm.test.AdminPageTest@392a04e7</td></tr>
 <tr>
-<td title='test.AdminPageTest.startExtent()'><b>startExtent</b><br>Test class: test.AdminPageTest</td>
+<td title='com.orangehrm.test.AdminPageTest.startExtent()'><b>startExtent</b><br>Test class: com.orangehrm.test.AdminPageTest</td>
 <td></td>
 <td>0</td>
-<td>test.AdminPageTest@392a04e7</td></tr>
+<td>com.orangehrm.test.AdminPageTest@392a04e7</td></tr>
 <tr>
-<td title='test.AdminPageTest.afterEachTestMethod()'><b>afterEachTestMethod</b><br>Test class: test.AdminPageTest</td>
+<td title='com.orangehrm.test.AdminPageTest.afterEachTestMethod()'><b>afterEachTestMethod</b><br>Test class: com.orangehrm.test.AdminPageTest</td>
 <td></td>
 <td>0</td>
-<td>test.AdminPageTest@392a04e7</td></tr>
+<td>com.orangehrm.test.AdminPageTest@392a04e7</td></tr>
 <tr>
-<td title='test.AdminPageTest.afterEachTestMethod()'><b>afterEachTestMethod</b><br>Test class: test.AdminPageTest</td>
+<td title='com.orangehrm.test.AdminPageTest.afterEachTestMethod()'><b>afterEachTestMethod</b><br>Test class: com.orangehrm.test.AdminPageTest</td>
 <td></td>
 <td>0</td>
-<td>test.AdminPageTest@392a04e7</td></tr>
+<td>com.orangehrm.test.AdminPageTest@392a04e7</td></tr>
 </table><p>
 <table width='100%' border='1' class='invocation-skipped'>
 <tr><td colspan='4' align='center'><b>SKIPPED TESTS</b></td></tr>
@@ -201,7 +201,7 @@ Command: [null, newSession {capabilities=[Capabilities {browserName: chrome, goo
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='test.AdminPageTest.AddJobTitle()'><b>AddJobTitle</b><br>Test class: test.AdminPageTest</td>
+<td title='com.orangehrm.test.AdminPageTest.AddJobTitle()'><b>AddJobTitle</b><br>Test class: com.orangehrm.test.AdminPageTest</td>
 <td><div><pre>org.openqa.selenium.SessionNotCreatedException: Could not start a new session. Response code 500. Message: session not created: This version of ChromeDriver only supports Chrome version 111
 Current browser version is 110.0.5481.178 with binary path C:\Program Files\Google\Chrome\Application\chrome.exe 
 Host info: host: &apos;KABIR&apos;, ip: &apos;192.168.1.158&apos;
@@ -281,9 +281,9 @@ Command: [null, newSession {capabilities=[Capabilities {browserName: chrome, goo
 	at com.intellij.rt.testng.RemoteTestNGStarter.main(RemoteTestNGStarter.java:109)
 </pre></div></td>
 <td>0</td>
-<td>test.AdminPageTest@392a04e7</td></tr>
+<td>com.orangehrm.test.AdminPageTest@392a04e7</td></tr>
 <tr>
-<td title='test.AdminPageTest.ChangePasswordCheckBox()'><b>ChangePasswordCheckBox</b><br>Test class: test.AdminPageTest</td>
+<td title='com.orangehrm.test.AdminPageTest.ChangePasswordCheckBox()'><b>ChangePasswordCheckBox</b><br>Test class: com.orangehrm.test.AdminPageTest</td>
 <td><div><pre>org.openqa.selenium.SessionNotCreatedException: Could not start a new session. Response code 500. Message: session not created: This version of ChromeDriver only supports Chrome version 111
 Current browser version is 110.0.5481.178 with binary path C:\Program Files\Google\Chrome\Application\chrome.exe 
 Host info: host: &apos;KABIR&apos;, ip: &apos;192.168.1.158&apos;
@@ -363,7 +363,7 @@ Command: [null, newSession {capabilities=[Capabilities {browserName: chrome, goo
 	at com.intellij.rt.testng.RemoteTestNGStarter.main(RemoteTestNGStarter.java:109)
 </pre></div></td>
 <td>0</td>
-<td>test.AdminPageTest@392a04e7</td></tr>
+<td>com.orangehrm.test.AdminPageTest@392a04e7</td></tr>
 </table><p>
 </body>
 </html>
