+++ test-output/testng-failed.xml
@@ -10,7 +10,7 @@
   <parameter name="url" value="https://opensource-demo.orangehrmlive.com/"/>
   <test thread-count="5" name="orangehrm tests(failed)">
     <classes>
-      <class name="test.AdminPageTest">
+      <class name="com.orangehrm.test.AdminPageTest">
         <methods>
           <include name="setUp"/>
           <include name="ChangePasswordCheckBox"/>
@@ -20,7 +20,7 @@
           <include name="generateReport"/>
           <include name="startExtent"/>
         </methods>
-      </class> <!-- test.AdminPageTest -->
+      </class> <!-- com.orangehrm.test.AdminPageTest -->
     </classes>
   </test> <!-- orangehrm tests(failed) -->
 </suite> <!-- Failed suite [orangehrm test] -->
